make Tab/Enter in search box work with the TW Locator
The Locator, if fully installed, adds tabs to the search box for Fields
and Locator, which prevent Tab and Enter from working at all if this
plugin is also installed. With this change, the Tab/Enter shortcuts will
work on the List (standard search) and Locator tabs, and TiddlyWiki core
behavior is retained on the Fields tab.